+++ DB/migrations/20210421135625-create-quality-of-life.js
@@ -12,25 +12,25 @@ module.exports = {
         type: Sequelize.STRING,
       },
       quality_of_life_index: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
       },
       purchasing_power_index: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
       },
       health_care_index: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
       },
       cost_of_living_index: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
       },
       property_price_to_income_ratio: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
       },
       pollution_index: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
       },
       climate_index: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
       },
       created_at: {
         allowNull: false,
